Move geometry.ts to shared location
