+++ mathesar_ui/src/geometry.ts
